Add scenario specific services
The AblaServiceProvider allows multiple ServiceProviders to be configured
with their own individual ServiceCollections. When a service is requested,
it will loop through each ServiceProvider in the Stack and attempt to
resolve the dependency.

If there is no registered service in any of the ServiceProviders within
the AlbaServiceProvider, an InvalidOperationException is thrown indicating
that the service could not be resolved.

By introducing this class and wrapping the ServiceProvider within the
scenario with the AlbaServiceProvider, it allows services to be specified
for a specific scenario and only for that scenario. This adds the
flexibility to provide alternative implementations for tests.